app/requirements.txt
pyspark==3.4.1
google-cloud-storage==2.14.0
google-auth==2.28.1
google-api-python-client==2.117.0
delta-spark
requests
yfinance